utility: Move ForwardModel out to its own module.
